requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
yt-dlp>=2024.4.9
openai-whisper>=20231117
ffmpeg-python>=0.2.0
spacy>=3.7.0
scikit-learn>=1.4.0
numpy>=1.26.0
httpx>=0.27.0
pytest>=8.1.0
pytest-asyncio>=0.23.0

# Optional — install for semantic search
# sentence-transformers>=2.7.0

# Optional — install for abstractive summarisation
# transformers>=4.40.0
# torch>=2.2.0
